WIP: make NSS tests work locally and in browser
