+++ e2e/node/nss-unauth.test.ts
@@ -19,10 +19,10 @@
 // SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-import { describe, expect, it, test } from "@jest/globals";
+import { describe, expect, it, test, beforeAll } from "@jest/globals";
 
 import { foaf, schema } from "rdf-namespaces";
-import { getNodeTestingEnvironment } from "@inrupt/internal-test-env";
+import { getNodeTestingEnvironment, getPodRoot } from "@inrupt/internal-test-env";
 import {
   getSolidDataset,
   setThing,
@@ -56,6 +56,8 @@ import {
   getBoolean,
   setBoolean,
 } from "../../src/index";
+import { Session } from "@inrupt/solid-client-authn-node";
+import { getNssSession } from "./getNssSession";
 
 const env = getNodeTestingEnvironment();
 if (env.environment !== "NSS") {
@@ -66,8 +68,25 @@ if (env.environment !== "NSS") {
 // This block of end-to-end tests should be removed once solid-client-authn-node works against NSS,
 // and the e2e tests have an NSS environment setup.
 describe("End-to-end tests with pre-existing data against resources in an NSS server", () => {
-  // TODO: Remove this hard coded URL and use the env variables
-  const rootContainer = "http://localhost:8443/publicWrite/";
+  let rootContainer: string;
+  let options: { fetch: typeof global.fetch };
+  let session: Session;
+  let pod: string;
+
+  beforeAll(async () => {
+    const owner = env.clientCredentials.owner;
+    if (owner.type  === "CSS Client Credentials") {
+      throw new Error("Unexpected CSS Credentials");
+    }
+    session = await getNssSession({
+      oidcIssuer: env.idp,
+      username: owner.id,
+      password: owner.secret,
+    });
+    pod = await getPodRoot(session);
+    rootContainer = `${pod}public${pod.startsWith('http://localhost:') ? 'Write' : ''}/`
+    options = { fetch: session.fetch };
+  }, 60_000);
 
   it("should be able to read and update data in a Pod", async () => {
     const randomNick = `Random nick ${Math.random()}`;
