+++ e2e/node/wac.test.ts
@@ -34,7 +34,6 @@ import {
   getPodRoot,
   getAuthenticatedSession,
 } from "@inrupt/internal-test-env";
-import fetchCookie from 'fetch-cookie';
 import {
   getSolidDatasetWithAcl,
   hasResourceAcl,
@@ -56,71 +55,13 @@ import {
   getAgentAccess as getAgentAccessUniversal,
   setPublicAccess as setPublicAccessUniversal,
 } from "../../src/access/universal";
+import { getNssSession } from "./getNssSession";
 
 const env = getNodeTestingEnvironment();
 const sessionResourcePrefix = "solid-client-tests/node/wac-";
 if (env.features?.WAC === "false" || env.features?.WAC === "") {
   // eslint-disable-next-line jest/no-focused-tests
   test.only(`Skipping unsupported WAC tests in ${env.environment}`, () => {});
-}
-
-const location = (res: Response) => typeof res.headers.get('location') === 'string'
-  ? new URL(res.headers.get('location')!, res.url)
-  : new URL(res.url);
-
-// Log into Node Solid Server v5.7.7
-export async function getNssSession(options: { oidcIssuer: string; username: string; password: string }) {
-  const session = new Session();
-  const url = await new Promise<string>(res => session.login({
-    oidcIssuer: options.oidcIssuer,
-    redirectUrl: 'http://localhost:3000/',
-    handleRedirect: res
-  }));
-
-  const cFetch = fetchCookie(fetch);
-  
-  const form = new URLSearchParams({
-    username: options.username,
-    password: options.password,
-    response_type: "code",
-    scope: "openid+offline_access+webid",
-    client_id: new URL(url).searchParams.get('client_id')!,
-    redirect_uri: "http://localhost:3000/",
-    state: new URL(url).searchParams.get('state')!
-  });
-
-  const fetchOptions: (() => RequestInit) = () => ({
-    method: 'POST',
-    headers: { 'content-type': 'application/x-www-form-urlencoded' },
-    body: form.toString(),
-    redirect: 'manual'
-  })
-
-  // Submit login details
-  let pwdRes = await cFetch(new URL('/login/password', options.oidcIssuer), fetchOptions());
-  let nextUrl: URL | false;
-
-  // Follow redirects. Terminate early if the next Location pointer contains the code.
-  while (pwdRes.status === 302 && !(nextUrl = location(pwdRes)).searchParams.has('code')) {
-    pwdRes = await cFetch(nextUrl, { redirect: 'manual' });
-  }
-
-  // Submit consent form if this is our first time logging into the given server
-  if (pwdRes.status !== 302) {
-    for (const mode of ['Read', 'Write', 'Append'])
-      form.append('access_mode', mode)
-    form.append('consent', 'true')
-
-    pwdRes = await cFetch(location(pwdRes).origin + location(pwdRes).pathname, fetchOptions());
-  }
-
-  // Follow redirects. Terminate early if the next Location pointer contains the code.
-  while (pwdRes.status === 302 && !(nextUrl = location(pwdRes)).searchParams.has('code')) {
-    pwdRes = await cFetch(nextUrl, { redirect: 'manual' });
-  }
-
-  await session.handleIncomingRedirect(location(pwdRes).toString());
-  return session;
 }
 
 describe("Authenticated end-to-end WAC", () => {
@@ -130,12 +71,15 @@ describe("Authenticated end-to-end WAC", () => {
   let pod: string;
 
   beforeEach(async () => {
+    const owner = env.clientCredentials.owner;
+    if (owner.type  === "CSS Client Credentials") {
+      throw new Error("Unexpected CSS Credentials");
+    }
     session = await getNssSession({
       oidcIssuer: env.idp,
-      username: "ownerName",
-      password: "ownerPassword1@",
+      username: owner.id,
+      password: owner.secret,
     });
-    // session = await getAuthenticatedSession(env);
     pod = await getPodRoot(session);
     sessionResource = `${pod}${sessionResourcePrefix}${session.info.sessionId}`;
     options = { fetch: session.fetch };
